requirements.txt
selenium
requests
opencv-python
pyautogui
pytest
pytest-rerunfailures
PYYAML
allure-pytest
Appium-Python-Client
openpyxl
xlrd
xlwt
xlutils
suds-jurko